"Assets in public directory cannot be imported from JavaScript."
fixed the error
```
Assets in public directory cannot be imported from JavaScript.
If you intend to import that asset, put the file in the src directory, and use /src/locales/en/translation.json instead of /public/locales/en/translation.json.
If you intend to use the URL of that asset, use /locales/en/translation.json?url.
Assets in public directory cannot be imported from JavaScript.
If you intend to import that asset, put the file in the src directory, and use /src/locales/fa/translation.json instead of /public/locales/fa/translation.json.
If you intend to use the URL of that asset, use /locales/fa/translation.json?url.
Assets in public directory cannot be imported from JavaScript.
If you intend to import that asset, put the file in the src directory, and use /src/12.png instead of /public/12.png.
If you intend to use the URL of that asset, use /12.png?url.
```
-added "i18next-http-backend": "^3.0.2" library
-instead of importing translation files it fetches them in runtime
-changed the importing path for '12.png' in LearningPathDetail.jsx from "../../public/12.png" to "../assets/12.png"